scripts/requirements.txt
opencv-python
asgiref
flask
torch
pillow
numpy
torchvision
uvicorn